Lazy sources and receivers
- Move interpolation out of sparse functions into its own module.
- Make interpolation behave lazy
